small update to break down file structure
