+++ web_app/app_factory.py
@@ -1,7 +1,7 @@
 from flask import Flask, render_template, request
 from user.view import user_bp
 from product.view import product_bp
-from models import Product
+from product.models import Product
 from decorator import login_required
 
 app = Flask(__name__, instance_relative_config=False)
